finish paySlips

+++ Auto1040.Api/Controllers/PaySlipController.cs
@@ -1,5 +1,6 @@
 ﻿using Auto1040.Api.PostModels;
 using Auto1040.Core.DTOs;
+using Auto1040.Core.Entities;
 using Auto1040.Core.Services;
 using AutoMapper;
 using Microsoft.AspNetCore.Authorization;
@@ -23,10 +24,12 @@ namespace Auto1040.Api.Controllers
             _mapper = mapper;
         }
 
+
         [HttpGet]
-        public ActionResult<IEnumerable<PaySlipDto>> GetAll()
+        public ActionResult<IEnumerable<PaySlipDto>> GetAll(int? userId)
         {
-            var result = _paySlipService.GetAllPaySlips();
+            var authId = GetUserId();
+            var result = _paySlipService.GetPaySlipByUserId(userId,authId);
             if (!result.IsSuccess)
                 return StatusCode(result.StatusCode, result.ErrorMessage);
 
@@ -36,6 +39,10 @@ namespace Auto1040.Api.Controllers
         [HttpGet("{id}")]
         public ActionResult<PaySlipDto> Get(int id)
         {
+            var userId = GetUserId();
+            if (!_paySlipService.IsPaySlipOwner(id, userId))
+                return Forbid("You are not authorized to access this resource.");
+
             var result = _paySlipService.GetPaySlipById(id);
             if (!result.IsSuccess)
                 return StatusCode(result.StatusCode, result.ErrorMessage);
@@ -43,48 +50,19 @@ namespace Auto1040.Api.Controllers
             return Ok(result.Data);
         }
 
-        [HttpPost("upload")]
-        [Consumes("multipart/form-data")]
-        public async Task<ActionResult<PaySlipDto>> Upload([FromForm] FileDto fileDto)
-        {
-            if (fileDto == null ||fileDto.File==null|| fileDto.File.Length == 0)
-            {
-                return BadRequest("No file uploaded");
-            }
-            var userId = GetUserId();
-            // Call the service to process the file and extract the necessary data
-            var result = await _paySlipService.ProcessPaySlipFileAsync(fileDto.File, userId);
-
-            if (!result.IsSuccess)
-            {
-                return StatusCode(result.StatusCode, result.ErrorMessage);
-            }
-
-            return Ok(result.Data);
-        }
-
-        [HttpPost]
-        public async Task<ActionResult<bool>> Add([FromBody] PaySlipPostModel paySlip)
-        {
-            if (paySlip == null)
-                return BadRequest("Pay slip data is required.");
-
-            var paySlipDto = _mapper.Map<PaySlipDto>(paySlip);
-            var result = await _paySlipService.AddPaySlipAsync(paySlipDto);
-            if (!result.IsSuccess)
-                return StatusCode(result.StatusCode, result.ErrorMessage);
-
-            return Ok(result.Data);
-        }
 
         [HttpPut("{id}")]
         public async Task<ActionResult<bool>> Update(int id, [FromBody] PaySlipPostModel paySlip)
         {
+            var userId = GetUserId();
+            if (!_paySlipService.IsPaySlipOwner(id, userId))
+                return Forbid("You are not authorized to access this resource.");
+
             if (paySlip == null)
                 return BadRequest("Pay slip data is required.");
 
             var paySlipDto = _mapper.Map<PaySlipDto>(paySlip);
-            var result =await _paySlipService.UpdatePaySlipAsync(id, paySlipDto);
+            var result = await _paySlipService.UpdatePaySlipAsync(id, paySlipDto);
             if (!result.IsSuccess)
                 return StatusCode(result.StatusCode, result.ErrorMessage);
 
@@ -94,11 +72,34 @@ namespace Auto1040.Api.Controllers
         [HttpDelete("{id}")]
         public ActionResult<bool> Delete(int id)
         {
+            var userId = GetUserId();
+            if (!_paySlipService.IsPaySlipOwner(id, userId))
+                return Forbid("You are not authorized to access this resource.");
+
             var result = _paySlipService.DeletePaySlip(id);
             if (!result.IsSuccess)
                 return StatusCode(result.StatusCode, result.ErrorMessage);
 
             return NoContent();
+        }
+
+        [HttpPost("upload")]
+        [Consumes("multipart/form-data")]
+        public async Task<ActionResult<PaySlipDto>> Upload([FromForm] FileDto fileDto)
+        {
+            if (fileDto == null || fileDto.File == null || fileDto.File.Length == 0)
+            {
+                return BadRequest("No file uploaded");
+            }
+            var userId = GetUserId();
+            var result = await _paySlipService.ProcessPaySlipFileAsync(fileDto.File, userId);
+
+            if (!result.IsSuccess)
+            {
+                return StatusCode(result.StatusCode, result.ErrorMessage);
+            }
+
+            return Ok(result.Data);
         }
 
         private int GetUserId()
@@ -110,5 +111,6 @@ namespace Auto1040.Api.Controllers
             }
             return int.Parse(userIdClaim);
         }
+
     }
 }
